feat: save debug status to local file
