+++ packages/vscode-extension/src/commandController.ts
@@ -1,20 +1,12 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { commands } from "vscode";
-
 import { FxError, Result, ok } from "@microsoft/teamsfx-api";
-
-import {
-  emptyProjectStatus,
-  getProjectStatus,
-  saveProjectStatus,
-} from "./chat/commands/nextstep/status";
-import { NecessaryActions } from "./chat/commands/nextstep/types";
+import { commands } from "vscode";
 import { workspaceUri } from "./globalVariables";
 import treeViewManager from "./treeview/treeViewManager";
 import { localize } from "./utils/localizeUtils";
-import { getFixedCommonProjectSettings } from "@microsoft/teamsfx-core";
+import { updateProjectStatus } from "./utils/projectStatusUtils";
 
 type CommandHandler = (...args: unknown[]) => Promise<Result<unknown, FxError>>;
 
@@ -87,19 +79,8 @@ class CommandController {
     const command = this.commandMap.get(commandName);
     if (command) {
       const result = await command.callback(...args);
-      const projectSettings = getFixedCommonProjectSettings(workspaceUri?.fsPath);
-      const p = projectSettings?.projectId ?? workspaceUri?.fsPath;
-      const actions = NecessaryActions.map((x) => x.toString());
-      if (p && actions.includes(commandName)) {
-        /// save project action running status
-        const status = (await getProjectStatus(p)) ?? emptyProjectStatus();
-        await saveProjectStatus(p, {
-          ...status,
-          [commandName]: {
-            result: result.isOk() ? "success" : "fail",
-            time: new Date(),
-          },
-        });
+      if (workspaceUri?.fsPath) {
+        await updateProjectStatus(workspaceUri.fsPath, commandName, result);
       }
       return result;
     }
